Разделил код на модули. Создал файлы data.js, util.js. Удалил файл function.js
